+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+
+### Fixed
+- Correct element location and size when using layout(), ignore(), strict(), content(), floating() and accessibility(). [Trello 1601](https://trello.com/c/mN1R9zVZ)
+
 ## [4.5.3] - 2020-02-28
 ### Fixed 
 - Removed deprecated Sonatype parent POM. [Trello 1556](https://trello.com/c/EGpks4Pv)